Feat: Llama API ChatBot 기능 빌드
- Llama ChatBot API 기능 구현
- INFO 레벨 로그 기능 구현
- Ollama 주소 환경변수 및 모델 이름 config 설정
- Llama 초기 설정 및 기본적인 예외 처리
- LlamaController 테스트 파일 추가
