+++ service/src/main/java/com/office/app/dto/ChatRequest.java
@@ -1,0 +1,12 @@
+package com.office.app.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ChatRequest {
+    private String prompt;
+}
