+++ service/src/main/java/com/office/app/dto/ChatResponse.java
@@ -1,0 +1,13 @@
+package com.office.app.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ChatResponse {
+    private String response;
+    private String status;
+}
